requirements.txt
attrs==17.4.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
daiquiri==1.3.0
idna==2.6
pyyaml==3.12
requests==2.18.4
urllib3==1.22
voluptuous==0.10.5
